fs: avoid adding non-thread-group task to LMK rbtree
To maintain the task adj RB tree, we add a task to the RB tree when fork,
and delete it when exit. The place is exactly the same as the linear
p->tasks list, say, nly when the task is thread_group_leader.

When task group_leader is changing, we make sure to add the new leader
into RB tree after its leader flag is set, task->exit_signal.

Change-Id: I8da47998510e531188feb067b491e92306be9414
Signed-off-by: Hong-Mei Li <a21834@motorola.com>
Signed-off-by: Alex Naidis <alex.naidis@linux.com>
Signed-off-by: Eliminater74 <eliminater74@gmail.com>