power: qpnp-fg: fix address and offset in fg_conventional_mem_write
Currently, in fg_conventional_mem_write, address is rounded off
to a multiple of 4 bytes and that is used to calculate offset.
This creates a problem when an address which is not a multiple
of 4 bytes is passed with offset 0. Data will be written to the
modified address with offset 0. Fix it by using the original
address to calculate both address and offset.

While at it, add a debug print to log when cycle counter is
stored and modify a debug print to pr_debug which is used less
often.

Change-Id: Ibb8088bd76ba658c718efe37ce2cd85d00f2fb4b
Signed-off-by: Subbaraman Narayanamurthy <subbaram@codeaurora.org>
Signed-off-by: Eliminater74 <eliminater74@gmail.com>