smb-lib: fix usbin_icl_change interrupt storm
We observed an interrupt storm on that interrupt
even after the usb path is suspended.

So keep this interrupt disabled as long as the usb
path is suspended.

Note that since the current limit votes can come in
prior to interrupt registration, rerun the election
on the votable after interrupts are registered.

Change-Id: Ib43b071b7b0decf8e50d27d585bea04b727f9475
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>
Signed-off-by: Eliminater74 <eliminater74@gmail.com>